cmake_modules/trike/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "trike"
description = "Simple doc comments for C++"
version = "0.0.1"
dependencies = [
  "libclang",
  "sphinx >= 8",
]
classifiers = [
  "Framework :: Sphinx :: Extension",
]

[project.optional-dependencies]
test = [
    "pytest >= 8",
]
